+++ addons/pt/saft/extensions.go
@@ -10,25 +10,14 @@ import (
 const (
 	ExtKeyExemption    cbc.Key = "pt-saft-exemption"
 	ExtKeyTaxRate      cbc.Key = "pt-saft-tax-rate"
-	ExtKeyInvoiceType  cbc.Key = "pt-saft-invoice-type"
 	ExtKeyProductType  cbc.Key = "pt-saft-product-type"
-	ExtKeyPaymentType  cbc.Key = "pt-saft-payment-type"
 	ExtKeyPaymentMeans cbc.Key = "pt-saft-payment-means"
-)
 
-// Invoice types
-const (
-	InvoiceTypeStandard       cbc.Code = "FT"
-	InvoiceTypeSimplified     cbc.Code = "FS"
-	InvoiceTypeInvoiceReceipt cbc.Code = "FR"
-	InvoiceTypeDebitNote      cbc.Code = "ND"
-	InvoiceTypeCreditNote     cbc.Code = "NC"
-)
-
-// Payment types
-const (
-	PaymentTypeCash  cbc.Code = "RC"
-	PaymentTypeOther cbc.Code = "RG"
+	// Document types extensions
+	ExtKeyInvoiceType  cbc.Key = "pt-saft-invoice-type"
+	ExtKeyWorkType     cbc.Key = "pt-saft-work-type"
+	ExtKeyMovementType cbc.Key = "pt-saft-movement-type"
+	ExtKeyPaymentType  cbc.Key = "pt-saft-payment-type"
 )
 
 // Tax rates
@@ -49,6 +38,39 @@ const (
 	ProductTypeFee     cbc.Code = "I"
 )
 
+// Document types
+const (
+	InvoiceTypeStandard       cbc.Code = "FT"
+	InvoiceTypeSimplified     cbc.Code = "FS"
+	InvoiceTypeInvoiceReceipt cbc.Code = "FR"
+	InvoiceTypeDebitNote      cbc.Code = "ND"
+	InvoiceTypeCreditNote     cbc.Code = "NC"
+
+	MovementTypeDeliveryNote cbc.Code = "GR"
+	MovementTypeWaybill      cbc.Code = "GT"
+	MovementTypeFixedAssets  cbc.Code = "GA"
+	MovementTypeConsignment  cbc.Code = "GC"
+	MovementTypeReturn       cbc.Code = "GD"
+
+	WorkTypeTableQueries      cbc.Code = "CM"
+	WorkTypeConsignmentCredit cbc.Code = "CC"
+	WorkTypeConsignmentInv    cbc.Code = "FC"
+	WorkTypeWorksheets        cbc.Code = "FO"
+	WorkTypePurchaseOrder     cbc.Code = "NE"
+	WorkTypeOther             cbc.Code = "OU"
+	WorkTypeBudgets           cbc.Code = "OR"
+	WorkTypeProforma          cbc.Code = "PF"
+	WorkTypeDocuments         cbc.Code = "DC"
+	WorkTypePremium           cbc.Code = "RP"
+	WorkTypeChargeback        cbc.Code = "RE"
+	WorkTypeCoInsurers        cbc.Code = "CS"
+	WorkTypeLeadCoInsurer     cbc.Code = "LD"
+	WorkTypeReinsurance       cbc.Code = "RA"
+
+	PaymentTypeCash  cbc.Code = "RC"
+	PaymentTypeOther cbc.Code = "RG"
+)
+
 var extensions = []*cbc.Definition{
 	{
 		Key: ExtKeyInvoiceType,
@@ -61,6 +83,16 @@ var extensions = []*cbc.Definition{
 				SAF-T's ~InvoiceType~ (Tipo de documento) specifies the type of a sales invoice. In GOBL,
 				this type can be set using the ~pt-saft-invoice-type~ extension in the tax section. GOBL
 				will set the extension for you based on the type and the tax tags you set in your invoice.
+
+				The table below shows how this mapping is done:
+
+				| Code | Name                | GOBL Type     | GOBL Tax Tag    |
+				| ---- | ------------------- | ------------- | --------------- |
+				| ~FT~ | Standard Invoice    | ~standard~    |                 |
+				| ~FS~ | Simplified Invoice  | ~standard~    | ~simplified~    |
+				| ~FR~ | Invoice-Receipt     | ~standard~    | ~invoice-receipt~ |
+				| ~ND~ | Debit Note          | ~debit-note~  |                 |
+				| ~NC~ | Credit Note         | ~credit-note~ |                 |
 
 				Example:
 
@@ -371,7 +403,7 @@ var extensions = []*cbc.Definition{
 			{
 				Code: "M15",
 				Name: i18n.String{
-					i18n.EN: "Margin scheme - Collector’s items and antiques / Decree-Law No. 199/96 of 18th October",
+					i18n.EN: "Margin scheme - Collector's items and antiques / Decree-Law No. 199/96 of 18th October",
 					i18n.PT: "Regime da margem de lucro - Objetos de coleção e antiguidades / Decreto-Lei n.° 199/96, de 18 de outubro",
 				},
 			},
@@ -561,6 +593,174 @@ var extensions = []*cbc.Definition{
 		},
 	},
 	{
+		Key: ExtKeyWorkType,
+		Name: i18n.String{
+			i18n.EN: "Document Type",
+			i18n.PT: "Tipo de documento",
+		},
+		Desc: i18n.String{
+			i18n.EN: here.Doc(`
+				SAF-T's ~WorkType~ (Tipo de documento de conferência) specifies the type of a working
+				document. In GOBL, this type can be set using the ~pt-saft-work-type~ extension in either
+				~bill.Invoice~ or ~bill.Order~ documents. GOBL will set the extension for you based on the
+				document type in some cases.
+
+				The table below shows the supported work type codes and their compatibility with GOBL objects:
+
+				| Code   | Name                            | GOBL Doc | GOBL Type  |
+				| ------ | ------------------------------- | -------- | ---------- |
+				| ~PF~   | Pró-forma                       | Invoice  | ~proforma~ |
+				| ~FC~   | Fatura de consignação           | Invoice  |            |
+				| ~CC~   | Credito de consignação          | Invoice  |            |
+				| ~CM~   | Consultas de mesa               | Order    |            |
+				| ~FO~   | Folhas de obra                  | Order    |            |
+				| ~NE~   | Nota de Encomenda               | Order    | ~purchase~ |
+				| ~OU~   | Outros                          | Order    |            |
+				| ~OR~   | Orçamentos                      | Order    | ~quote~    |
+				| ~DC~   | Documentos de conferência       | Order    |            |
+				| ~RP~   | Prémio ou recibo de prémio      | Order    |            |
+				| ~RE~   | Estorno ou recibo de estorno    | Order    |            |
+				| ~CS~   | Imputação a co-seguradoras      | Order    |            |
+				| ~LD~   | Imputação a co-seguradora líder | Order    |            |
+				| ~RA~   | Resseguro aceite                | Order    |            |
+
+				Example for a proforma invoice:
+
+				~~~js
+				{
+					"$schema": "https://gobl.org/draft-0/bill/invoice",
+					"type": "proforma",
+					// ...
+					"tax": {
+						"ext": {
+							"pt-saft-work-type": "PF"
+						}
+					},
+					// ...
+				~~~
+
+				Example for a purchase order:
+
+				~~~js
+				{
+					"$schema": "https://gobl.org/draft-0/bill/order",
+					"type": "purchase",
+					// ...
+					"tax": {
+						"ext": {
+							"pt-saft-work-type": "NE"
+						}
+					},
+					// ...
+				~~~
+			`),
+		},
+		Values: []*cbc.Definition{
+			{
+				Code: WorkTypeTableQueries,
+				Name: i18n.String{
+					i18n.EN: "Table orders",
+					i18n.PT: "Consultas de mesa",
+				},
+			},
+			{
+				Code: WorkTypeConsignmentCredit,
+				Name: i18n.String{
+					i18n.EN: "Consignment credit note",
+					i18n.PT: "Credito de consignação",
+				},
+			},
+			{
+				Code: WorkTypeConsignmentInv,
+				Name: i18n.String{
+					i18n.EN: "VAT-compliant consignment invoice (Article 38)",
+					i18n.PT: "Fatura de consignação nos termos do art.º 38º do código do IVA",
+				},
+			},
+			{
+				Code: WorkTypeWorksheets,
+				Name: i18n.String{
+					i18n.EN: "Work orders",
+					i18n.PT: "Folhas de obra",
+				},
+			},
+			{
+				Code: WorkTypePurchaseOrder,
+				Name: i18n.String{
+					i18n.EN: "Purchase order",
+					i18n.PT: "Nota de Encomenda",
+				},
+			},
+			{
+				Code: WorkTypeOther,
+				Name: i18n.String{
+					i18n.EN: "Other documents",
+					i18n.PT: "Outros",
+				},
+			},
+			{
+				Code: WorkTypeBudgets,
+				Name: i18n.String{
+					i18n.EN: "Quotations",
+					i18n.PT: "Orçamentos",
+				},
+			},
+			{
+				Code: WorkTypeProforma,
+				Name: i18n.String{
+					i18n.EN: "Pro forma invoice",
+					i18n.PT: "Pró-forma",
+				},
+			},
+			{
+				Code: WorkTypeDocuments,
+				Name: i18n.String{
+					i18n.EN: "Delivery verification documents",
+					i18n.PT: "Documentos emitidos que sejam suscetíveis de apresentação ao cliente para conferência de mercadorias ou de prestação de serviços",
+				},
+				Desc: i18n.String{
+					i18n.EN: "For data up to 2017-06-30",
+					i18n.PT: "Para dados até 2017-06-30",
+				},
+			},
+			{
+				Code: WorkTypePremium,
+				Name: i18n.String{
+					i18n.EN: "Premium Receipt",
+					i18n.PT: "Prémio ou recibo de prémio",
+				},
+			},
+			{
+				Code: WorkTypeChargeback,
+				Name: i18n.String{
+					i18n.EN: "Chargeback Receipt",
+					i18n.PT: "Estorno ou recibo de estorno",
+				},
+			},
+			{
+				Code: WorkTypeCoInsurers,
+				Name: i18n.String{
+					i18n.EN: "Co-insurers Allocation",
+					i18n.PT: "Imputação a co-seguradoras",
+				},
+			},
+			{
+				Code: WorkTypeLeadCoInsurer,
+				Name: i18n.String{
+					i18n.EN: "Lead Co-insurer Allocation",
+					i18n.PT: "Imputação a co-seguradora líder",
+				},
+			},
+			{
+				Code: WorkTypeReinsurance,
+				Name: i18n.String{
+					i18n.EN: "Accepted Reinsurance",
+					i18n.PT: "Resseguro aceite",
+				},
+			},
+		},
+	},
+	{
 		Key: ExtKeyPaymentMeans,
 		Name: i18n.String{
 			i18n.EN: "Payment Means",
@@ -570,7 +770,7 @@ var extensions = []*cbc.Definition{
 			i18n.EN: here.Doc(`
 				The SAF-T's ~PaymentMechanism~ (Meios de pagamento) field specifies the payment means in a
 				sales invoice or payment. GOBL provides the ~pt-saft-payment-means~ extension to set this
-				value in your ~bill.Invoice~ advances or in you ~bill.Receipt~ method. GOBL maps certain
+				value in your ~bill.Invoice~ advances or in you ~bill.Payment~ method. GOBL maps certain
 				payment mean keys automatically to this extension:
 
 				| Code | Name                                               | GOBL Payment Means                                    |
@@ -730,6 +930,83 @@ var extensions = []*cbc.Definition{
 				Name: i18n.String{
 					i18n.EN: "Supplementary compensation",
 					i18n.PT: "Títulos de compensação extrassalarial",
+				},
+			},
+		},
+	},
+	{
+		Key: ExtKeyMovementType,
+		Name: i18n.String{
+			i18n.EN: "Movement Type",
+			i18n.PT: "Tipo de documento",
+		},
+		Desc: i18n.String{
+			i18n.EN: here.Doc(`
+				SAF-T's ~MovementType~ (Tipo de documento de movimentação de mercadorias) specifies the type of
+				a delivery document. In GOBL,this type can be set using the ~pt-saft-movement-type~ extension.
+				If not provided explicitly, GOBL will set the extension for you based on the type of your delivery
+				document.
+
+				The table below shows how this mapping is done:
+
+				| Code | Name                | GOBL Type     |
+				| ---- | ------------------- | ------------- |
+				| ~GR~ | Delivery note       | ~note~        |
+				| ~GT~ | Waybill             | ~waybill~     |
+
+				Example:
+
+				~~~js
+				{
+					"$schema": "https://gobl.org/draft-0/bill/delivery",
+					// ...
+					"type": "note",
+					// ...
+					"ext": {
+						"pt-saft-movement-type": "GR"
+					},
+					// ...
+				~~~
+			`),
+		},
+		Values: []*cbc.Definition{
+			{
+				Code: MovementTypeDeliveryNote,
+				Name: i18n.String{
+					i18n.EN: "Delivery note",
+					i18n.PT: "Guia de remessa",
+				},
+			},
+			{
+				Code: MovementTypeWaybill,
+				Name: i18n.String{
+					i18n.EN: "Waybill",
+					i18n.PT: "Guia de transporte",
+				},
+				Desc: i18n.String{
+					i18n.EN: "Include global waybills here",
+					i18n.PT: "Incluir aqui as guias globais",
+				},
+			},
+			{
+				Code: MovementTypeFixedAssets,
+				Name: i18n.String{
+					i18n.EN: "Guide to the movement own fixed assets",
+					i18n.PT: "Guia de movimentação de ativos fixos próprios",
+				},
+			},
+			{
+				Code: MovementTypeConsignment,
+				Name: i18n.String{
+					i18n.EN: "Consignment note",
+					i18n.PT: "Guia de consignação",
+				},
+			},
+			{
+				Code: MovementTypeReturn,
+				Name: i18n.String{
+					i18n.EN: "Returns slip or note",
+					i18n.PT: "Guia ou nota de devolução",
 				},
 			},
 		},